Db work2

+++ app/models/generateFillTables/tblVolAssoc.php
@@ -20,19 +20,20 @@ function createTableVolAssoc($conn)
 {
     // asociere intre asociatii si voluntari
     if (pg_query($conn, 'CREATE TABLE tblVolAssoc(
-    vol_id INTEGER NOT NULL,
-    assoc_id INTEGER NOT NULL,
-    active BOOLEAN NOT NULL DEFAULT TRUE,      /* volunteer - association relationship is active? When a volunteer stops working
-                                                    for an association (active set to false), its history will be preserved */
-    hours_worked INTEGER,
-    bonus_points INTEGER,                      /* little hearts received for small tasks */
-    rating_id serial UNIQUE,                   /* used as foreign key in tblFeedback */
-    created_on TIMESTAMP NOT NULL,
-    PRIMARY KEY (vol_id, assoc_id),
-    CONSTRAINT volassoc_vol_fkey FOREIGN KEY (vol_id)
-    REFERENCES tblVolunteers(id),
-    CONSTRAINT volassoc_assoc_fkey FOREIGN KEY (assoc_id)
-    REFERENCES tblAssociations(id)
+        id serial UNIQUE,
+        vol_id INTEGER NOT NULL, /*fkey*/
+        assoc_id INTEGER NOT NULL, /*fkey*/
+        active BOOLEAN NOT NULL DEFAULT TRUE,      /* volunteer - association relationship is active? When a volunteer stops working
+                                                        for an association (active set to false), its history will be preserved */
+        -- hours_worked INTEGER,
+        -- bonus_points INTEGER,                      /* little hearts received for small tasks */
+        /*rating_id serial UNIQUE,                    used as foreign key in tblFeedback */
+        created_on TIMESTAMP NOT NULL,
+        PRIMARY KEY (vol_id, assoc_id),
+        CONSTRAINT volassoc_vol_fkey FOREIGN KEY (vol_id)
+        REFERENCES tblVolunteers(id),
+        CONSTRAINT volassoc_assoc_fkey FOREIGN KEY (assoc_id)
+        REFERENCES tblAssociations(id)
     )  ')) {
         echo "Table VolAssoc created!<br>";
     } else {
@@ -62,12 +63,15 @@ function insert_VolAssoc($conn, $assoc, $vol)
 {
     if (rand(0, 100) < 9) {                     /* 9% probabilitate ca voluntarul $vol sa lucreze la asociatia $assoc */
         $query  ='INSERT INTO tblVolAssoc 
-        (assoc_id, vol_id, hours_worked, bonus_points, created_on) VALUES 
-        (' . $assoc . ',' . $vol .', ' . strval(rand(0, 500)) .', ' . strval(rand(0, 150)) .', current_timestamp)';
+        (assoc_id, vol_id, created_on) VALUES 
+        (' . $assoc . ',' . $vol .', current_timestamp)';
 
-        if (pg_query($conn, $query)) {
-            echo "Record added in tblVolAssoc";
+        try {
+            if (pg_query($conn, $query)) {
+                echo "Record added in tblVolAssoc";
+            }
+        } catch (Exception $e) {
+            // echo $e->getMessage();
         }
     }
 }
-
